Extract jq related schema and function
